% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Perform Principal Component Analysis (PCA)}
\usage{
PCA(X, k)
}
\arguments{
\item{X}{A numeric matrix data or a data frame representing the input data}

\item{k}{An integer specifying the number of principal components to retain}
}
\value{
A list that has the mean vector, orthogonal matrix, Factor scores,and Eigenvalues
}
\description{
This function performs Principal Component Analysis (PCA) on the input data matrix X
and returns the projected data onto the selected principal components
}
\examples{
#Example dataset
test.dat <- matrix(rnorm(500), nrow=100, ncol=5)
result.PCA <- PCA(test.dat, k = 2)
print(result.PCA)
PCA

}
