% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Perform Principal Component Analysis (PCA)}
\usage{
PCA(X, k)
}
\arguments{
\item{X}{A numeric matrix data or a data frame representing the input data}

\item{k}{An integer specifying the number of principal components to retain}
}
\value{
A matrix containing the projected data onto the selected principal components
}
\description{
This function performs Principal Component Analysis (PCA) on the input data matrix X
and returns the projected data onto the selected principal components
}
\examples{
#Example dataset
PCA.dat <- matrix(c(2, 4, 1, 3, 5, 7, 6, 8, 9), nrow = 3, ncol = 3, byrow = TRUE)
test.dat <- as.numeric(PCA.dat)
result.PCA <- PCA(test.dat, k = 2)
print(result.PCA)
PCA

}
